+++ database/Cargo.toml
@@ -11,9 +11,11 @@ scylla = "0.9.0"
 tracing = "0.1.34"
 uuid = { version = "1.3.0", optional = true }
 
-near-primitives = "0.17.0"
 readnode-primitives = { path = "../readnode-primitives" }
+
+near-primitives = "0.17.0"
 near-crypto = "0.17.0"
 
 [features]
 scylla_db_tracing = ["dep:prettytable-rs", "dep:uuid"]
+account_access_keys = []
