+++ readnode-primitives/Cargo.toml
@@ -6,10 +6,10 @@ edition = "2021"
 [dependencies]
 anyhow = "1.0.70"
 borsh = "0.10.2"
-serde = { version = "1.0.145", features = ["derive"] }
-serde_json = "1.0.85"
 num-bigint = "0.3"
 num-traits = "0.2.15"
+serde = { version = "1.0.145", features = ["derive"] }
+serde_json = "1.0.85"
 
 near-indexer-primitives = "0.17.0"
 near-primitives = "0.17.0"
