+++ rpc-server/Cargo.toml
@@ -70,4 +70,4 @@ submit_tx_methods = []
 tracing-instrumentation = []
 scylla_db_tracing = ["database/scylla_db_tracing"]
 shadow_data_consistency = ["dep:assert-json-diff"]
-account_access_keys = []
+account_access_keys = ["database/account_access_keys"]
