+++ rpc-server/src/config.rs
@@ -1,5 +1,7 @@
 use crate::modules::blocks::CacheBlock;
+use crate::storage;
 use clap::Parser;
+use database::ScyllaStorageManager;
 
 #[derive(Parser)]
 #[clap(author, version, about, long_about = None)]
@@ -111,11 +113,25 @@ impl Opts {
             .build()
             .expect("Failed to build LakeConfig"))
     }
+
+    pub async fn prepare_db_manager(&self) -> anyhow::Result<impl database::RpcDbManager> {
+        let db_manager = *storage::ScyllaDBManager::new(
+            &self.scylla_url,
+            self.scylla_user.as_deref(),
+            self.scylla_password.as_deref(),
+            self.scylla_preferred_dc.as_deref(),
+            Some(self.scylla_keepalive_interval),
+            self.max_retry,
+            self.strict_mode,
+        )
+        .await?;
+        Ok(db_manager)
+    }
 }
 
 pub struct ServerContext {
     pub s3_client: near_lake_framework::s3_fetchers::LakeS3Client,
-    pub scylla_db_manager: std::sync::Arc<Box<dyn database::BaseDbManager + Sync + Send + 'static>>,
+    pub db_manager: std::sync::Arc<Box<dyn database::RpcDbManager + Sync + Send + 'static>>,
     pub near_rpc_client: near_jsonrpc_client::JsonRpcClient,
     pub s3_bucket_name: String,
     pub genesis_config: near_chain_configs::GenesisConfig,
@@ -130,23 +146,28 @@ pub struct ServerContext {
 }
 
 impl ServerContext {
+    #[allow(clippy::too_many_arguments)]
     pub fn new(
         s3_client: near_lake_framework::s3_fetchers::LakeS3Client,
-        db_manager: impl database::BaseDbManager + Sync + Send + 'static,
+        db_manager: impl database::RpcDbManager + Sync + Send + 'static,
         near_rpc_client: near_jsonrpc_client::JsonRpcClient,
         s3_bucket_name: String,
         genesis_config: near_chain_configs::GenesisConfig,
-        blocks_cache: std::sync::Arc<std::sync::RwLock<crate::cache::LruMemoryCache<u64, CacheBlock>>>,
+        blocks_cache: std::sync::Arc<
+            std::sync::RwLock<crate::cache::LruMemoryCache<u64, CacheBlock>>,
+        >,
         final_block_height: std::sync::Arc<std::sync::atomic::AtomicU64>,
         compiled_contract_code_cache: std::sync::Arc<CompiledCodeCache>,
         contract_code_cache: std::sync::Arc<
-            std::sync::RwLock<crate::cache::LruMemoryCache<near_primitives::hash::CryptoHash, Vec<u8>>>,
+            std::sync::RwLock<
+                crate::cache::LruMemoryCache<near_primitives::hash::CryptoHash, Vec<u8>>,
+            >,
         >,
         max_gas_burnt: near_primitives_core::types::Gas,
     ) -> Self {
         Self {
             s3_client,
-            scylla_db_manager: std::sync::Arc::new(Box::new(db_manager)),
+            db_manager: std::sync::Arc::new(Box::new(db_manager)),
             near_rpc_client,
             s3_bucket_name,
             genesis_config,
@@ -157,7 +178,6 @@ impl ServerContext {
             max_gas_burnt,
         }
     }
-
 }
 pub struct CompiledCodeCache {
     pub local_cache: std::sync::Arc<
